Fixed Problem Where Column Name is Missing First Letter
Removed code initially included when data was fetched via TCP rather than with the HTTP wrapper. The TCP version relied on cutting out the page header, and when this code was applied to the data retrieved with the HTTP wrapper, the first 1-2 characters were being removed from the dataset (on the column header row). Removing the "substring" code works fine as the HTTP wrapper knows when to start reading the data.